# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ogis: Open Graph Images as a Service",
    "description": "Generate Open Graph images dynamically via URL parameters",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://img.ogis.dev"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "image"
        ],
        "operationId": "generate",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Title text for the image",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "Description text for the image",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtitle",
            "in": "query",
            "description": "Subtitle text (above title)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logo",
            "in": "query",
            "description": "Optional logo image URL",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "image",
            "in": "query",
            "description": "Optional custom image URL",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "template",
            "in": "query",
            "description": "Template name (e.g., \"twilight\", \"daybreak\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signature",
            "in": "query",
            "description": "HMAC signature (required when authentication is enabled)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Output format: png (default), webp, jpeg",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scale",
            "in": "query",
            "description": "Scale factor for output resolution (0.1-max_scale, default: 1.0)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "description": "Quality for lossy formats (1-100, default: 90, ignored for PNG)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "extra",
            "in": "query",
            "description": "Additional parameters (used for color customization)",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              },
              "propertyNames": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully generated PNG image (1200x630)",
            "content": {
              "image/png": {}
            }
          },
          "400": {
            "description": "Invalid input - field exceeds maximum length or invalid hex color"
          },
          "401": {
            "description": "Authentication required - missing signature"
          },
          "403": {
            "description": "Forbidden - invalid signature or SSRF blocked"
          },
          "404": {
            "description": "Template not found"
          },
          "422": {
            "description": "Unprocessable - invalid image URL, unsupported format, or image too large"
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "Bad gateway - upstream image fetch failed"
          },
          "503": {
            "description": "Service unavailable - server overloaded"
          },
          "504": {
            "description": "Gateway timeout - image fetch timed out"
          }
        }
      }
    },
    "/c/{blob}": {
      "get": {
        "tags": [
          "image"
        ],
        "operationId": "generate_compressed",
        "parameters": [
          {
            "name": "blob",
            "in": "path",
            "description": "Base64url-encoded compressed parameter blob, produced by an ogis SDK",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated image. Content-Type and dimensions depend on the encoded format/scale: image/png at 1200x630 by default, or image/jpeg / image/webp when the blob specifies.",
            "content": {
              "image/png": {}
            }
          },
          "400": {
            "description": "Invalid compressed URL - malformed blob or unsupported version"
          },
          "401": {
            "description": "Authentication required - missing or invalid signature"
          },
          "403": {
            "description": "Forbidden - SSRF blocked (private IP)"
          },
          "422": {
            "description": "Unprocessable - invalid image URL, unsupported format, or image too large"
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "Bad gateway - upstream image fetch failed"
          },
          "503": {
            "description": "Service unavailable - server overloaded"
          },
          "504": {
            "description": "Gateway timeout - image fetch timed out"
          }
        }
      }
    },
    "/c/{blob}/{sig}": {
      "get": {
        "tags": [
          "image"
        ],
        "operationId": "generate_compressed_signed",
        "parameters": [
          {
            "name": "blob",
            "in": "path",
            "description": "Base64url-encoded compressed parameter blob, produced by an ogis SDK",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sig",
            "in": "path",
            "description": "base64url-nopad of the leading 6 bytes of HMAC-SHA256 over [version byte ++ uncompressed body] (8 characters)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated image. Content-Type and dimensions depend on the encoded format/scale: image/png at 1200x630 by default, or image/jpeg / image/webp when the blob specifies.",
            "content": {
              "image/png": {}
            }
          },
          "400": {
            "description": "Invalid compressed URL - malformed blob or unsupported version"
          },
          "401": {
            "description": "Authentication required - missing or invalid signature"
          },
          "403": {
            "description": "Forbidden - SSRF blocked (private IP)"
          },
          "422": {
            "description": "Unprocessable - invalid image URL, unsupported format, or image too large"
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "Bad gateway - upstream image fetch failed"
          },
          "503": {
            "description": "Service unavailable - server overloaded"
          },
          "504": {
            "description": "Gateway timeout - image fetch timed out"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Health check endpoint",
        "description": "Returns a simple \"ok\" response to indicate the service is running",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/templates": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "List all registered template names.",
        "description": "Useful for benchmarking and tooling that needs to enumerate every template\nthe running server can generate without coupling to the on-disk layout.",
        "operationId": "list_templates",
        "responses": {
          "200": {
            "description": "List of template names",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatesResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "description": "JSON error response body",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable error codes for programmatic handling",
        "enum": [
          "VALIDATION_FIELD_TOO_LONG",
          "VALIDATION_INVALID_URL",
          "VALIDATION_INVALID_HEX_COLOR",
          "VALIDATION_INVALID_FORMAT",
          "VALIDATION_INVALID_SCALE",
          "VALIDATION_INVALID_QUALITY",
          "AUTH_INVALID_SIGNATURE_FORMAT",
          "INVALID_COMPRESSED_URL",
          "AUTH_MISSING_SIGNATURE",
          "AUTH_INVALID_SIGNATURE",
          "SSRF_PRIVATE_IP_BLOCKED",
          "TEMPLATE_NOT_FOUND",
          "IMAGE_TOO_LARGE",
          "IMAGE_INVALID_CONTENT_TYPE",
          "IMAGE_HTTP_NOT_ALLOWED",
          "RENDER_FAILED",
          "SVG_PARSE_FAILED",
          "IMAGE_ENCODE_FAILED",
          "FONT_ERROR",
          "INTERNAL_ERROR",
          "UPSTREAM_FETCH_FAILED",
          "SERVICE_OVERLOADED",
          "UPSTREAM_TIMEOUT"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "description": "Error detail within the response",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable error code"
          },
          "details": {
            "type": [
              "string",
              "null"
            ],
            "description": "Additional details about the error"
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "Field that caused the error (if applicable)"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
      "OgParams": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Description text for the image"
          },
          "format": {
            "type": [
              "string",
              "null"
            ],
            "description": "Output format: png (default), webp, jpeg",
            "example": "webp"
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional custom image URL"
          },
          "logo": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional logo image URL"
          },
          "quality": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Quality for lossy formats (1-100, default: 90, ignored for PNG)",
            "example": 90,
            "minimum": 0
          },
          "scale": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Scale factor for output resolution (0.1-max_scale, default: 1.0)",
            "example": 0.5
          },
          "signature": {
            "type": [
              "string",
              "null"
            ],
            "description": "HMAC signature (required when authentication is enabled)",
            "example": "a1b2c3d4e5f6..."
          },
          "subtitle": {
            "type": [
              "string",
              "null"
            ],
            "description": "Subtitle text (above title)"
          },
          "template": {
            "type": [
              "string",
              "null"
            ],
            "description": "Template name (e.g., \"twilight\", \"daybreak\")"
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Title text for the image"
          }
        },
        "additionalProperties": {
          "type": "string",
          "description": "Additional parameters (used for color customization)"
        }
      },
      "TemplatesResponse": {
        "type": "object",
        "required": [
          "default",
          "templates"
        ],
        "properties": {
          "default": {
            "type": "string",
            "description": "Name of the template returned when the request supplies no `template` param."
          },
          "templates": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "All template names registered at startup (file-based + auto-composed gradients)."
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "image",
      "description": "Image generation endpoints"
    },
    {
      "name": "monitoring",
      "description": "Service monitoring and health checks"
    }
  ]
}